Change Dockerfile to support not patched version
The patched version is only available to paid users. I'm using this on a
development environment using the free version of ODSEE available on
[Oracle Cloud](https://edelivery.oracle.com/osdc/faces/SoftwareDelivery)

Has the same zip from the patched version. `dsee7.zip` so I changed the
script to only try to download the patched version if this file is not
already available.